requirements.txt
streamlit>=1.28.0
ccxt>=4.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
plotly>=5.18.0
matplotlib>=3.8.0
pandas-ta>=0.3.14b0
python-binance>=1.0.19
scipy>=1.12.0
requests>=2.31.0